requirements.txt
folium==0.15.0
geopy==2.4.1
matplotlib==3.8.2
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.1.0
numpy==1.24.3
googlemaps==4.10.0 